+++ lib/test/cases/multiline_splits.py
@@ -1,0 +1,9 @@
+"Some simple string"
+
+"Some simple string split over \
+two lines\
+with backslashes"
+
+f" \
+Single quoted f-string split over {'multiple'} lines\
+with backslashes"