+++ lib/test/snapshots/multiline_splits_py312.py
@@ -1,0 +1,8 @@
+f'{
+    'f-string expression split over multiple lines'
+}'
+
+f'{
+    'f-string expression with a string split\
+    over multiple lines'
+}'